Now have support for loading in our "inspection" JSON documents
This is setting up for a tool to migrate our existing data files to
a new database (possibly Mongo). In particular, we want to do this
from Haskell code and so for the first time must support loading
in the final JSON documents produced in the past.

Because of custom JSON instancing for Google Places data, this
required some fancy Haskell Orphan Instance trickery along with
some code refactorization to keep these instances apart.