Fixed a bug in parsing of the (naive) dates in inspections
The dates are simply mm/dd/yyyy but are assumed to be localized
to EST/EDT. Our parsing code was interpreting them as UTC always,
leading to occasional screw-ups when converting them to the local
zome. We need UTC because that's the native format for BSON and
MongoDB.

Took the opportunity to add unit testing for this project as well,
to test the `Inspection.parseDate` function.