Attempt to fix date parsing weirdness with daylight savings
I'm wondering now if the error was really in the unit test --
introducing its own problem by converting the parsed UTCTime into
the local zone+dst sometimes.

In any case, changed the Inspection parsing code to use the LocalTime
constructor with midnight, force it to always be 00:00:00. And then
am testing the values of the Y M and D explicitly in the unit test.

Also, added an inverse test for date parsing, parsing a bad date
string should fail.

Finally, the type of parseDate is now Either String UTCTime to
represent possible failure. This was ugly partial business before,
sloppy work!