Fixed a bug with parsing date strings and daylight-savings time
We were using the time zone info for right now to figure out the
zone, even when the date we're parsing is from a part of the year
with a different daylight-savings status than now. This was causing
off-by-one-hour problems with parsing older dates.

This problem manifests itself when you try to parse inspections
from the past (or I suppose the future).

The date of inspection is now used to daylight-savings-localize
the zone info.

Also added this to the unit tests, they were not testing for this
problem.