Changed Status and IssueType To/FromBSON instancing
I was having problems using these types as MongoDB filter criteria
because they couldn't easily be converted into BSON DocumentS. This
solution, to turn each one into an entire Document with a Label is
sub-optimal, but is the best we can do.

This change did not have any effect on the (hacky) To/FromBSON
marshalling being done for whole Feedback data structures,
unfortunately.

The reasons for this are a fundamental structure mismatch between
BSON and JSON. Valid JSON values are recursive and can be primitives
in addition to Object for containment of anything else:

      data Value =
         Object !Object  
         Array !Array    
         String !Text    
         Number !Scientific    
         Bool !Bool   
         Null

BSON was designed in a less flexible way where the only
representation of a Document is roughly like this:

      type Document = [Field]
      data Field = !Label := Value
      ...

Closes Kitchen Snitch #137